Add LRUcache test
